Transition to generic relation api base checker
